feat: refine durability verification, integrity sweep, degraded-mode policy, add API layer and supply-chain trust cards
- connector/durability.rs: Refactored durability violation detection
  to use structured error types and improved threshold comparison
  logic for tier-based integrity checks.
- connector/mod.rs: Registers new durability sub-module exports.
- policy/integrity_sweep_scheduler.rs: Minor fix to scheduling
  interval boundary condition.
- security/degraded_mode_policy.rs: Fix to degraded-state signaling
  predicate.
- api/: New API surface layer providing external-facing endpoints
  for node health, trust status, and policy queries.
- supply_chain/trust_card.rs: Implements trust cards — portable,
  signed attestations of a component's provenance, build environment,
  and dependency graph, used for supply-chain integrity verification.

+++ crates/franken-node/src/policy/integrity_sweep_scheduler.rs
@@ -732,7 +732,7 @@ mod tests {
         let mut intervals = Vec::new();
 
         for i in 1..=5 {
-            let evidence = EvidenceTrajectory::new(i * 2, i as u32, 0.9 - (i as f64 * 0.1), Trend::Degrading, i);
+            let evidence = EvidenceTrajectory::new(i * 2, i as u32, 0.9 - (i as f64 * 0.1), Trend::Degrading, i as u64);
             sched.update_trajectory(&evidence);
             intervals.push(sched.next_sweep_interval().as_millis());
         }
